refactor: use ApiVersion enum throughout HTTP/WS gateway and split V1/V2 routers
Replace raw version_prefix: &str with ApiVersion enum across all HTTP
gateway and path handler functions. This ensures version routing uses a
single source of truth (ApiVersion::prefix()) instead of scattered string
literals.

- Add ApiVersion::prefix() method and document its relationship with
  DelegateApiVersion (routing vs WASM execution versioning)
- Split http_gateway/v1.rs into separate V1 and V2 router files, each
  exposing a routes() function merged in the parent
- Inline path_handlers/v1.rs into path_handlers.rs (eliminate submodule)
- Eliminate websocket_commands_v1/v2 wrappers using per-route
  Extension<ApiVersion> on nested routers
- Add V2 path rewriting test and error path test for get_file_path
